update ProductBrowser
move inputRef from parent components to ProductBrowser itself, change
way of making input focus after submit, update test
